LUN-121: Fix for random order
